[fix] fixed update_action_processor in case of nonexisting skin on depot

+++ opsiclientd/EventProcessing.py
@@ -555,7 +555,8 @@ class EventProcessingThread(KillableThread, ServiceConnection): # pylint: disabl
 		if RUNNING_ON_WINDOWS:
 			# saving current opsi-script skin (set during opsi-client-agent setup with optional corporate identity)
 			if os.path.exists(os.path.join(actionProcessorLocalDir, "skin")):
-				shutil.rmtree(os.path.join(actionProcessorLocalTmpDir, "skin"))
+				if os.path.exists(os.path.join(actionProcessorLocalTmpDir, "skin")):
+					shutil.rmtree(os.path.join(actionProcessorLocalTmpDir, "skin"))
 				shutil.move(os.path.join(actionProcessorLocalDir, "skin"), os.path.join(actionProcessorLocalTmpDir, "skin"))
 
 		if not os.path.exists(os.path.join(actionProcessorLocalTmpDir, actionProcessorFilename)):
